move makeSchemaFrom functions in separate file
